+++ src/mutate.rs
@@ -211,7 +211,7 @@ mod test {
         let tool = CargoTool::new();
         let source_tree = tool.find_root(tree_path).unwrap();
         let options = Options::default();
-        let mutants = walk_tree(&tool, &source_tree, &options, &Console::new())
+        let mutants = walk_tree(&tool, &source_tree, &[], &options, &Console::new())
             .unwrap()
             .mutants;
         assert_eq!(mutants.len(), 3);
@@ -265,9 +265,15 @@ mod test {
         let tree_path = Utf8Path::new("testdata/tree/hang_avoided_by_attr");
         let tool = CargoTool::new();
         let source_tree = tool.find_root(tree_path).unwrap();
-        let mutants = walk_tree(&tool, &source_tree, &Options::default(), &Console::new())
-            .unwrap()
-            .mutants;
+        let mutants = walk_tree(
+            &tool,
+            &source_tree,
+            &[],
+            &Options::default(),
+            &Console::new(),
+        )
+        .unwrap()
+        .mutants;
         let descriptions = mutants.iter().map(Mutant::describe_change).collect_vec();
         insta::assert_snapshot!(
             descriptions.join("\n"),
@@ -280,9 +286,15 @@ mod test {
         let tree_path = Utf8Path::new("testdata/tree/factorial");
         let tool = CargoTool::new();
         let source_tree = tool.find_root(tree_path).unwrap();
-        let mutants = walk_tree(&tool, &source_tree, &Options::default(), &Console::new())
-            .unwrap()
-            .mutants;
+        let mutants = walk_tree(
+            &tool,
+            &source_tree,
+            &[],
+            &Options::default(),
+            &Console::new(),
+        )
+        .unwrap()
+        .mutants;
         assert_eq!(mutants.len(), 3);
 
         let mut mutated_code = mutants[0].mutated_code();